Create 'c40Encode()' and 'c40Decode()' as module imports.
