+++ modules/EventsMRVA.js
@@ -6,6 +6,8 @@ import { DigitalSeal } from "./icao9303/DigitalSeal.js";
 import { DigitalSealV4 } from "./icao9303/DigitalSealV4.js";
 import { DEFAULT_PHOTO, DEFAULT_SIGNATURE_IMAGE } from "./icao9303/utilities/default-images.js";
 import { generateMRZCheckDigit } from "./icao9303/utilities/generate-mrz-check-digit.js";
+import { c40Encode } from "./icao9303/utilities/c40-encode.js";
+import { c40Decode } from "./icao9303/utilities/c40-decode.js";
 
 /**
  * `EventsMRVA` describes an ALFA Furry Events Visa in the MRV-A document size
@@ -369,7 +371,7 @@ class EventsMRVA {
    */
   set passportNumber(value) {
     this.#document.passportNumber = value;
-    this.#seal.features.set(0x05, DigitalSeal.c40Encode(value));
+    this.#seal.features.set(0x05, c40Encode(value));
     this.#setDigitalSealMRZ();
   }
 
@@ -633,7 +635,7 @@ class EventsMRVA {
    *     setting any properties shown on the MRZ.
    */
   #setDigitalSealMRZ() {
-    this.#seal.features.set(0x01, DigitalSeal.c40Encode(
+    this.#seal.features.set(0x01, c40Encode(
       this.mrzLine1 + this.mrzLine2.slice(0, 28)
     ));
   }
@@ -644,7 +646,7 @@ class EventsMRVA {
    */
   #setAllValuesFromDigitalSeal() {
     const TWO_DIGIT_YEAR_START = 32;
-    const SEAL_MRZ = DigitalSeal.c40Decode(this.#seal.features.get(0x01));
+    const SEAL_MRZ = c40Decode(this.#seal.features.get(0x01));
     if (SEAL_MRZ[53] !== generateMRZCheckDigit(
       SEAL_MRZ.slice(44, 53).replace(/ /gi, "<")
     )) {
@@ -697,7 +699,7 @@ class EventsMRVA {
       this.#document.numberOfEntries = this.#seal.features.get(0x03)[0];
     }
     this.#document.passportNumber =
-        DigitalSeal.c40Decode(this.#seal.features.get(0x05));
+        c40Decode(this.#seal.features.get(0x05));
   }
 }
