+++ modules/EventsPassport.js
@@ -7,6 +7,8 @@ import { DigitalSealV4 } from "./icao9303/DigitalSealV4.js";
 import { DEFAULT_PHOTO, DEFAULT_SIGNATURE_IMAGE } from "./icao9303/utilities/default-images.js";
 import { dateToVIZ } from "./icao9303/utilities/date-to-viz.js";
 import { generateMRZCheckDigit } from "./icao9303/utilities/generate-mrz-check-digit.js";
+import { c40Encode } from "./icao9303/utilities/c40-encode.js";
+import { c40Decode } from "./icao9303/utilities/c40-decode.js";
 
 /**
  * `EventsPassport` describes an ALFA Furry Events Passport, a TD3-sized
@@ -551,7 +553,7 @@ class EventsPassport {
    *     setting any properties shown on the MRZ.
    */
   #setDigitalSealMRZ() {
-    this.#seal.features.set(0x01, DigitalSeal.c40Encode(
+    this.#seal.features.set(0x01, c40Encode(
       this.mrzLine1 + this.mrzLine2.slice(0, 28)
     ));
   }
@@ -562,7 +564,7 @@ class EventsPassport {
    */
   #setAllValuesFromDigitalSeal() {
     const TWO_DIGIT_YEAR_START = 32;
-    const SEAL_MRZ = DigitalSeal.c40Decode(this.#seal.features.get(0x01));
+    const SEAL_MRZ = c40Decode(this.#seal.features.get(0x01));
     if (SEAL_MRZ[53] !== generateMRZCheckDigit(
       SEAL_MRZ.slice(44, 53).replace(/ /gi, "<")
     )) {
