+++ modules/icao9303/DigitalSeal.js
@@ -247,14 +247,6 @@ class DigitalSeal {
     return value.slice(start);
   }
 
-  static #c40SHIFT1 = Symbol("c40SHIFT1");
-  /**
-   * A filler character used when C40 encoding requires one more byte to
-   *     complete a triple.
-   * @readonly
-   */
-  static get c40SHIFT1() { return this.#c40SHIFT1; }
-
   /**
    * A magic constant used to identify a byte array as a VDS.
    * @readonly
@@ -266,479 +258,6 @@ class DigitalSeal {
    * @readonly
    */
   static get signatureMarker() { return 0xFF; }
-
-  /**
-   * Encode a string using the C40 encoding scheme defined in ICAO 9303 part 13,
-   *     which adds the use of the filler character '<' in the encoding scheme,
-   *     into a number array.
-   * @param { string } string - A string containing only A-Z, 0-9, <SPACE>, and
-   *     the symbol '<'.
-   * @example
-   * // Returns [235, 4, 102, 169]
-   * DigitalSeal.c40Encode("XK CD");
-   */
-  static c40Encode(string) {
-    const output = [];
-    let i16;
-    let u1 = null;
-    let u2 = null;
-    let u3 = null;
-    [...string].forEach((character, i) => {
-      if (u1 === null && ((string.length - (i + 1)) === 0)) {
-        output.push(0xFE);
-        output.push(DigitalSeal.#charToDataMatrixASCII(character));
-      } else if (u1 === null) {
-        u1 = DigitalSeal.#charToC40(character);
-      } else if (u2 === null) {
-        u2 = DigitalSeal.#charToC40(character);
-        if ((string.length - (i + 1)) === 0) {
-          u3 = DigitalSeal.#charToC40(DigitalSeal.c40SHIFT1);
-          i16 = (1600 * u1) + (40 * u2) + u3 + 1;
-          output.push(Math.floor(i16 / 256));
-          output.push(i16 % 256);
-          i16 = null;
-          u1 = null;
-          u2 = null;
-          u3 = null;
-        }
-      } else {
-        u3 = DigitalSeal.#charToC40(character);
-        i16 = (1660 * u1) + (40 * u2) + u3 + 1;
-        output.push(Math.floor(i16 / 256));
-        output.push(i16 % 256);
-        i16 = null;
-        u1 = null;
-        u2 = null;
-        u3 = null;
-      }
-    });
-    return output;
-  }
-
-  /**
-   * Decode a C40-encoded number array as defined in ICAO 9303 part 13 into a
-   *     string. 
-   * @param { number[] } c40
-   * @example
-   * // Returns "XKCD"
-   * DigitalSeal.c40Decode([235, 17, 254, 69]);
-   */
-  static c40Decode(c40) {
-    let output = "";
-    let i1 = null;
-    let i2 = null;
-    c40.forEach((byte) => {
-      if (i1 === null) {
-        i1 = byte;
-      } else {
-        if (i1 === parseInt("0xFE")) {
-          output += DigitalSeal.#dataMatrixASCIIToChar(byte);
-        } else {
-          i2 = byte;
-          const I16 = (i1 * 256) + i2;
-          const U1 = Math.floor((I16 - 1) / 1600);
-          const U2 = Math.floor((I16 - (U1 * 1600) - 1) / 40);
-          const U3 = I16 - (U1 * 1600) - (U2 * 40) - 1;
-          output += DigitalSeal.#c40ToChar(U1);
-          output += DigitalSeal.#c40ToChar(U2);
-          if (DigitalSeal.#c40ToChar(U3) !== DigitalSeal.c40SHIFT1) {
-            output += DigitalSeal.#c40ToChar(U3);
-          }
-        }
-        i1 = null;
-        i2 = null;
-      }
-    });
-    return output;
-  }
-
-  /**
-   * Get the corresponding C40 value for a character.
-   * @param { string | DigitalSeal.c40SHIFT1 } char - A-Z, 0-9, <SPACE>, the
-   *     symbol '<', or `DigitalSeal.c40SHIFT1`.
-   * @example
-   * // Returns 17
-   * DigitalSeal.#charToC40("D");
-   */
-  static #charToC40(char) {
-    if (char !== DigitalSeal.c40SHIFT1) {
-      if (char.length !== 1) {
-        throw new RangeError(
-          "Input must be a string of one character or SHIFT1 " +
-              "(DigitalSeal.c40SHIFT1)."
-        );
-      }
-    }
-    if (char === DigitalSeal.c40SHIFT1) {
-      return 0;
-    }
-    else {
-      switch (char.toUpperCase()) {
-        case " ":
-        case "<":
-          return 3;
-        case "0":
-          return 4;
-        case "1":
-          return 5;
-        case "2":
-          return 6;
-        case "3":
-          return 7;
-        case "4":
-          return 8;
-        case "5":
-          return 9;
-        case "6":
-          return 10;
-        case "7":
-          return 11;
-        case "8":
-          return 12;
-        case "9":
-          return 13;
-        case "A":
-          return 14;
-        case "B":
-          return 15;
-        case "C":
-          return 16;
-        case "D":
-          return 17;
-        case "E":
-          return 18;
-        case "F":
-          return 19;
-        case "G":
-          return 20;
-        case "H":
-          return 21;
-        case "I":
-          return 22;
-        case "J":
-          return 23;
-        case "K":
-          return 24;
-        case "L":
-          return 25;
-        case "M":
-          return 26;
-        case "N":
-          return 27;
-        case "O":
-          return 28;
-        case "P":
-          return 29;
-        case "Q":
-          return 30;
-        case "R":
-          return 31;
-        case "S":
-          return 32;
-        case "T":
-          return 33;
-        case "U":
-          return 34;
-        case "V":
-          return 35;
-        case "W":
-          return 36;
-        case "X":
-          return 37;
-        case "Y":
-          return 38;
-        case "Z":
-          return 39;
-        default:
-          throw new TypeError(
-            "Characters in Digital Seals may only contain the characters 0-9," +
-                "A-Z, <SPACE>, the symbol '<', or SHIFT1 " +
-                "(DigitalSeal.c40SHIFT1)."
-          );
-      }
-    }
-  }
-
-  /**
-   * Get the corresponding character for a C40 value.
-   * @param { number } char - A number in the range of 0-39.
-   * @example
-   * // Returns "D"
-   * DigitalSeal.#c40ToChar(17);
-   */
-  static #c40ToChar(char) {
-    switch (char) {
-      case 0:
-        return DigitalSeal.c40SHIFT1;
-      case 3:
-        return " ";
-      case 4:
-        return "0";
-      case 5:
-        return "1";
-      case 6:
-        return "2";
-      case 7:
-        return "3";
-      case 8:
-        return "4";
-      case 9:
-        return "5";
-      case 10:
-        return "6";
-      case 11:
-        return "7";
-      case 12:
-        return "8";
-      case 13:
-        return "9";
-      case 14:
-        return "A";
-      case 15:
-        return "B";
-      case 16:
-        return "C";
-      case 17:
-        return "D";
-      case 18:
-        return "E";
-      case 19:
-        return "F";
-      case 20:
-        return "G";
-      case 21:
-        return "H";
-      case 22:
-        return "I";
-      case 23:
-        return "J";
-      case 24:
-        return "K";
-      case 25:
-        return "L";
-      case 26:
-        return "M";
-      case 27:
-        return "N";
-      case 28:
-        return "O";
-      case 29:
-        return "P";
-      case 30:
-        return "Q";
-      case 31:
-        return "R";
-      case 32:
-        return "S";
-      case 33:
-        return "T";
-      case 34:
-        return "U";
-      case 35:
-        return "V";
-      case 36:
-        return "W";
-      case 37:
-        return "X";
-      case 38:
-        return "Y";
-      case 39:
-        return "Z";
-      default:
-        throw new RangeError(
-          "A C40 value in Digital Seals must be 0 or in the range of 3-39."
-        );
-    }
-  }
-
-  /**
-   * Get the corresponding DataMatrix ASCII value for a character.
-   * @param { string } char - A-Z, 0-9, <SPACE>, or the symbol '<'.
-   * @example
-   * // Returns 69
-   * DigitalSeal.#charToDataMatrixASCII("D");
-   */
-  static #charToDataMatrixASCII(char) {
-    if (char.length !== 1) {
-      throw new RangeError(
-        "Input must be one character containing 0-9, A-Z, <SPACE>, or the " +
-            "symbol '<'."
-      );
-    }
-    switch (char.toUpperCase()) {
-      case " ":
-      case "<":
-        return 33;
-      case "0":
-        return 49;
-      case "1":
-        return 50;
-      case "2":
-        return 51;
-      case "3":
-        return 52;
-      case "4":
-        return 53;
-      case "5":
-        return 54;
-      case "6":
-        return 55;
-      case "7":
-        return 56;
-      case "8":
-        return 57;
-      case "9":
-        return 58;
-      case "A":
-        return 66;
-      case "B":
-        return 67;
-      case "C":
-        return 68;
-      case "D":
-        return 69;
-      case "E":
-        return 70;
-      case "F":
-        return 71;
-      case "G":
-        return 72;
-      case "H":
-        return 73;
-      case "I":
-        return 74;
-      case "J":
-        return 75;
-      case "K":
-        return 76;
-      case "L":
-        return 77;
-      case "M":
-        return 78;
-      case "N":
-        return 79;
-      case "O":
-        return 80;
-      case "P":
-        return 81;
-      case "Q":
-        return 82;
-      case "R":
-        return 83;
-      case "S":
-        return 84;
-      case "T":
-        return 85;
-      case "U":
-        return 86;
-      case "V":
-        return 87;
-      case "W":
-        return 88;
-      case "X":
-        return 89;
-      case "Y":
-        return 90;
-      case "Z":
-        return 91;
-      default:
-        throw new TypeError(
-          "Characters in Digital Seals may only contain the characters 0-9, " +
-              "A-Z, <SPACE>, or the symbol '<'."
-        );
-    }
-  }
-
-  /**
-   * Get the corresponding character for a DataMatrix ASCII value.
-   * @param { number } char - 33, 49-58, or 66-91.
-   * @example
-   * // Returns "D"
-   * DigitalSeal.#dataMatrixASCIIToChar(69);
-   */
-  static #dataMatrixASCIIToChar(char) {
-    switch (char) {
-      case 33:
-        return " ";
-      case 49:
-        return "0";
-      case 50:
-        return "1";
-      case 51:
-        return "2";
-      case 52:
-        return "3";
-      case 53:
-        return "4";
-      case 54:
-        return "5";
-      case 55:
-        return "6";
-      case 56:
-        return "7";
-      case 57:
-        return "8";
-      case 58:
-        return "9";
-      case 66:
-        return "A";
-      case 67:
-        return "B";
-      case 68:
-        return "C";
-      case 69:
-        return "D";
-      case 70:
-        return "E";
-      case 71:
-        return "F";
-      case 72:
-        return "G";
-      case 73:
-        return "H";
-      case 74:
-        return "I";
-      case 75:
-        return "J";
-      case 76:
-        return "K";
-      case 77:
-        return "L";
-      case 78:
-        return "M";
-      case 79:
-        return "N";
-      case 80:
-        return "O";
-      case 81:
-        return "P";
-      case 82:
-        return "Q";
-      case 83:
-        return "R";
-      case 84:
-        return "S";
-      case 85:
-        return "T";
-      case 86:
-        return "U";
-      case 87:
-        return "V";
-      case 88:
-        return "W";
-      case 89:
-        return "X";
-      case 90:
-        return "Y";
-      case 91:
-        return "Z";
-      default:
-        throw new RangeError(
-          "A DataMatrix ASCII value in Digital Seals must be 33, 49-58, or " +
-              "66-91."
-        );
-    }
-  }
 
   /**
    * Get a corresponding byte array for a calendar date.
