+++ modules/icao9303/utilities/c40shift1.js
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2023 Don Geronimo <https://sentamal.in/>
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+/**
+ * A filler character used when C40 encoding requires one more byte to complete
+ *     a triple.
+ * @readonly
+ */
+export const C40SHIFT1 = Symbol("C40SHIFT1");
